Update Secure PGP keys and Yubikey NEO.md
Step 5. of signing instructions is slightly unclear. Initially I thought it meant I should upload the key to a keyserver. Then I realised that it was in the context of air-gapped machines as mentioned in step 1, and simply meant moving the signed key from the air-gapped machine back to the network connected machine.